Remove all namespaces from the controllers. This is done because rails routes do not play nice with the namespaces in controller classnames.
